Revert "KPN update"
