+++ assets/schema/ensemble_schema.json
@@ -1177,12 +1177,6 @@
           "type": "string",
           "description": "ID to be referenced later"
         },
-        "maxWidth": {
-          "type": "integer",
-          "description": "The max width of this Input widget (default 700)",
-          "minimum": 0,
-          "maximum": 5000
-        },
         "label": {
           "type": "string",
           "description": "Label for your widget"
@@ -1211,68 +1205,7 @@
         },
         "styles": {
           "allOf": [
-            { "$ref": "#/$defs/baseStyles" },
-            {
-              "type": "object",
-              "properties": {
-                "variant": {
-                  "type": "string",
-                  "description": "Select a pre-defined look and feel for this Input widget. This property can be defined in the theme to apply to all Input widgets.",
-                  "oneOf": [
-                    {
-                      "const": "underline",
-                      "description": "draw an Underline below this input widget. This property can be defined in the theme to apply to all Input widgets."
-                    },
-                    {
-                      "const": "box",
-                      "description": "draw a Box border around this input widget. This property can be defined in the theme to apply to all Input widgets."
-                    }
-                  ]
-                },
-                "contentPadding": {
-                  "type": ["integer", "string"],
-                  "description": "Padding around your input content with CSS-style notation e.g. margin: 5 20 5"
-                },
-                "fillColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The fill color for this input fields. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "borderRadius": {
-                  "type": "integer",
-                  "minimum": 0,
-                  "description": "The border radius for this Input widget. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "borderWidth": {
-                  "type": "integer",
-                  "minimum": 0,
-                  "description": "The border width for this Input widget. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "borderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The base border color for this input widget. This border color determines the look and feel of your input, while the other colors are overrides for different states. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "enabledBorderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The border color for this input field if enabled. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "disabledBorderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The border color when this input field is disabled. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "errorBorderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The border color when there are errors on this input field. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "focusedBorderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The border color when this input field is receiving focus. This property can be defined in the theme to apply to all Input widgets."
-                },
-                "focusedErrorBorderColor": {
-                  "$ref": "#/$defs/typeColors",
-                  "description": "The border color of this input field when it is receiving focus in its error state. This property can be defined in the theme to apply to all Input widgets."
-                }
-              }
-            }
+            { "$ref": "#/$defs/baseStyles" }
           ]
         }
       }
