--- ensemble/integration_tests/themedApp/theme.ensemble
@@ -1,6 +1,0 @@
-Widgets:
-  Input:
-    # default variant is Underline
-    borderRadius: 5
-    borderColor: 0xFF404040
-    focusedBorderColor: 0xFF49DEFF
