+++ lib/util/utils.dart
@@ -665,12 +665,8 @@ class Utils {
 
   // temporary workaround for internal translation so we dont have to duplicate the translation files in all repos
   static String translateWithFallback(String key, String fallback) {
-    if (Utils.globalAppKey.currentContext != null) {
-      String output = FlutterI18n.translate(
-          Utils.globalAppKey.currentContext!, key);
-      return output != key ? output : fallback;
-    }
-    return fallback;
+    String output = FlutterI18n.translate(Utils.globalAppKey.currentContext!, key);
+    return output != key ? output : fallback;
   }
   // explicitly return null if we can't find the translation key
   static String? translateOrNull(String key) {
