+++ lib/widget/form_time.dart
@@ -4,8 +4,7 @@ import 'package:ensemble/framework/event.dart';
 import 'package:ensemble/page_model.dart';
 import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/utils.dart';
-import 'package:ensemble/widget/input/form_helper.dart';
-import 'package:ensemble/widget/helpers/widgets.dart';
+import 'package:ensemble/widget/form_helper.dart';
 import 'package:ensemble/widget/widget_registry.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
 import 'package:flutter/material.dart';
@@ -62,32 +61,35 @@ class TimeController extends FormFieldController {
 class TimeState extends FormFieldWidgetState<Time> {
   @override
   Widget buildWidget(BuildContext context) {
-    return InputWrapper(
-        controller: widget.controller,
-        widget: FormField<DateTime>(
-            key: validatorKey,
-            validator: (value) {
-              if (widget._controller.required &&
-                  widget._controller.value == null) {
-                return Utils.translateWithFallback(
-                    'ensemble.input.required', 'This field is required');
-              }
-              return null;
-            },
-            builder: (FormFieldState<DateTime> field) {
-              return InputDecorator(
-                  decoration: inputDecoration.copyWith(
-                      border: InputBorder.none,
-                      enabledBorder: InputBorder.none,
-                      disabledBorder: InputBorder.none,
-                      focusedBorder: InputBorder.none,
-                      errorText: field.errorText),
-                  child: Row(mainAxisSize: MainAxisSize.min, children: [
-                    InkWell(
-                        child: nowBuildWidget(),
-                        onTap: isEnabled() ? () => _selectTime(context) : null)
-                  ]));
-            }));
+    return FormField<DateTime>(
+      key: validatorKey,
+      validator: (value) {
+        if (widget._controller.required && widget._controller.value == null) {
+          return Utils.translateWithFallback('ensemble.input.required', 'This field is required');
+        }
+        return null;
+      },
+      builder: (FormFieldState<DateTime> field) {
+        return InputDecorator(
+          decoration: inputDecoration.copyWith(
+            border: InputBorder.none,
+            enabledBorder: InputBorder.none,
+            disabledBorder: InputBorder.none,
+            focusedBorder: InputBorder.none,
+            errorText: field.errorText
+          ),
+          child: Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              InkWell(
+                child: nowBuildWidget(),
+                onTap: isEnabled() ? () => _selectTime(context) : null
+              )
+            ]
+          )
+        );
+      }
+    );
 
   }
