+++ lib/widget/helpers/widgets.dart
@@ -1,10 +1,8 @@
 /// This class contains common widgets for use with Ensemble widgets.
 
-import 'package:ensemble/widget/input/form_helper.dart';
 import 'package:ensemble/widget/helpers/controllers.dart';
-import 'package:ensemble/framework/theme/theme_manager.dart';
+import 'package:ensemble/widget/helpers/theme_manager.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
 
 /// wraps around a widget and gives it common box attributes
 class BoxWrapper extends StatelessWidget {
@@ -70,105 +68,28 @@ class BoxWrapper extends StatelessWidget {
                   ? GradientBorder(
                     gradient: boxController.borderGradient!,
                     width: boxController.borderWidth?.toDouble() ??
-                          ThemeManager().getBorderThickness(context)
+                          ThemeManager.getBorderThickness(context)
                   )
                   : Border.all(
                       color: boxController.borderColor ??
-                          ThemeManager().getBorderColor(context),
+                          ThemeManager.getBorderColor(context),
                       width: boxController.borderWidth?.toDouble() ??
-                          ThemeManager().getBorderThickness(context)),
+                          ThemeManager.getBorderThickness(context)),
               borderRadius: boxController.borderRadius?.getValue(),
               boxShadow: !boxController.hasBoxShadow()
                   ? null
                   : <BoxShadow>[
                       BoxShadow(
                         color: boxController.shadowColor ??
-                            ThemeManager().getShadowColor(context),
+                            ThemeManager.getShadowColor(context),
                         blurRadius:
                             boxController.shadowRadius?.toDouble() ??
-                                ThemeManager().getShadowRadius(context),
+                                ThemeManager.getShadowRadius(context),
                         offset: boxController.shadowOffset ??
-                            ThemeManager().getShadowOffset(context),
+                            ThemeManager.getShadowOffset(context),
                         blurStyle: boxController.shadowStyle ??
-                            ThemeManager().getShadowStyle(context))
+                            ThemeManager.getShadowStyle(context))
                     ]));
-  }
-
-}
-
-/// wrap the input widget (which stretches 100% to its parent) to guard against
-/// the case where it is put inside a Row without expanded flag.
-class InputWrapper extends StatelessWidget {
-  const InputWrapper({
-    super.key,
-    required this.widget,
-    required this.controller
-  });
-  final Widget widget;
-  final FormFieldController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(builder: (context, constraints) {
-      int? maxWidth = controller.maxWidth;
-
-      // inside a e.g. Row but not wrapping inside Expanded.
-      // This is the error condition we need to fix
-      if (!constraints.hasBoundedWidth && !controller.expanded) {
-        // this is our workaround for the layout error when the user use these
-        // input widgets (which stretch to their parent) inside a Row (which allow
-        // its child to have as much space as it wants) without using expanded flag.
-        // We'll cap their max width so the user can adjust it
-        // instead of seeing a blank screen
-        maxWidth ??= 700;
-      }
-
-      return maxWidth == null
-          ? widget
-          : ConstrainedBox(
-              constraints: BoxConstraints(maxWidth: maxWidth.toDouble()),
-              child: widget);
-    });
-  }
-  
-}
-
-
-
-/// Display a Text content followed by a clear icon.
-/// Clicking the icon will invoke the callback.
-/// Note that this is a StatelessWidget, so clearing out the value
-/// is the responsibility of the parent widget who uses this.
-class ClearableInput extends StatelessWidget {
-  const ClearableInput({
-    super.key,
-    required this.text,
-    required this.onCleared,
-    this.textStyle
-  });
-  final String text;
-  final TextStyle? textStyle;
-  final dynamic onCleared;
-
-  @override
-  Widget build(BuildContext context) {
-    if (text.isEmpty) {
-      return const SizedBox.shrink();
-    }
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Flexible(
-          child: Text(text, maxLines: 1, style: textStyle)
-        ),
-        const SizedBox(width: 4),
-        InkWell(
-          onTap: onCleared,
-          child: const Icon(
-              Icons.close,
-              size: 20))
-      ]
-    );
   }
 
 }
