+++ lib/widget/markdown.dart
@@ -6,7 +6,7 @@ import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/utils.dart';
 import 'package:ensemble/framework/widget/widget.dart' as framework;
 import 'package:ensemble/widget/helpers/controllers.dart';
-import 'package:ensemble/framework/theme/theme_manager.dart';
+import 'package:ensemble/widget/helpers/theme_manager.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/material.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
@@ -68,7 +68,7 @@ class MarkdownState extends framework.WidgetState<Markdown> {
     MarkdownStyleSheet styles = MarkdownStyleSheet.fromTheme(Theme.of(context)).copyWith(
       p: widget._controller.textStyle,
       a: widget._controller.linkStyle??TextStyle(
-        color: ThemeManager().getPrimaryColor(context)
+        color: ThemeManager.getPrimaryColor(context)
       ),
     );
