+++ lib/widget/weeklyscheduler.dart
@@ -1,6 +1,6 @@
 import 'package:ensemble/framework/widget/widget.dart';
 import 'package:ensemble/util/utils.dart';
-import 'package:ensemble/widget/input/form_time.dart';
+import 'package:ensemble/widget/form_time.dart';
 import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:ensemble/widget/widget_util.dart' as widget_utils;
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
