+++ lib/widget/widget_registry.dart
@@ -17,13 +17,13 @@ import 'package:ensemble/widget/button.dart';
 import 'package:ensemble/widget/carousel.dart';
 import 'package:ensemble/widget/chart_highcharts_builder.dart';
 import 'package:ensemble/widget/divider.dart';
-import 'package:ensemble/widget/input/dropdown.dart';
+import 'package:ensemble/widget/dropdown.dart';
 import 'package:ensemble/widget/ensemble_icon.dart';
-import 'package:ensemble/widget/input/form_checkbox.dart';
-import 'package:ensemble/widget/input/form_date.dart';
+import 'package:ensemble/widget/form_checkbox.dart';
+import 'package:ensemble/widget/form_date.dart';
 import 'package:ensemble/widget/form_daterange.dart';
-import 'package:ensemble/widget/input/form_textfield.dart';
-import 'package:ensemble/widget/input/form_time.dart';
+import 'package:ensemble/widget/form_textfield.dart';
+import 'package:ensemble/widget/form_time.dart';
 import 'package:ensemble/widget/image.dart';
 import 'package:ensemble/widget/lottie/lottie.dart';
 import 'package:ensemble/widget/map.dart';
