+++ test/widget/checkbox_switch_test.dart
@@ -1,5 +1,5 @@
-import 'package:ensemble/widget/input/form_checkbox.dart';
-import 'package:ensemble/widget/input/form_date.dart';
+import 'package:ensemble/widget/form_checkbox.dart';
+import 'package:ensemble/widget/form_date.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/intl.dart';
